Optimize indexer and cache
